requirements.txt
streamlit
requests
python-dotenv
fpdf
PyMuPDF
